Package delete pop up

+++ src/app/master/components/package/package.component.ts
@@ -8,6 +8,8 @@ import { Package } from '../../model/package.class';
 import { PackagePopupComponent } from '../package-popup/package-popup.component';
 import Swal from 'sweetalert2';
 import { ToastrService } from 'ngx-toastr';
+import { InstituteserviceService } from '../../service/instituteservice.service';
+import { SubscriptionService } from '../../service/subscription.service';
 
 @Component({
   selector: 'app-package',
@@ -28,15 +30,20 @@ export class PackageComponent implements OnInit {
   @ViewChild(MatPaginator) paginator: MatPaginator;
   @ViewChild(MatSort) sort: MatSort;
   packageData: any;
+  public instituteDetails = [];
+  public subscriptionList = [];
 
   constructor(
     private packageService: PackageService,
     public dialog: MatDialog,
-    private toastr: ToastrService
+    private toastr: ToastrService,
+    public instituteService: InstituteserviceService,
+    private subscriptionService: SubscriptionService
   ) {}
 
   ngOnInit(): void {
     this.loadPackageData();
+    this.getAllDataOfSubscription();
   }
 
   loadPackageData() {
@@ -62,6 +69,22 @@ export class PackageComponent implements OnInit {
     );
   }
 
+  getAllInstituteData() {
+    this.instituteService.getAllInstitute().subscribe(
+      (response: any) => {
+        this.instituteDetails = response.result;
+      },
+      (error) => {
+        console.error('Not data get', error);
+      }
+    );
+  }
+
+  getAllDataOfSubscription() {
+    this.subscriptionService.getAllSubscriptions().subscribe((data: any) => {
+      this.subscriptionList = data.result;
+    });
+  }
 
   openAddPackagePopup() {
     let dialogBoxSettings = {
@@ -103,29 +126,42 @@ export class PackageComponent implements OnInit {
   }
 
   deletePackage(element: Package) {
-    Swal.fire({
-      title: 'Are you sure?',
-      text: 'You are about to delete this package.',
-      icon: 'warning',
-      showCancelButton: true,
-      confirmButtonText: 'Delete',
-      cancelButtonText: 'Cancel',
-    }).then((result) => {
-      if (result.isConfirmed) {
-        this.packageService.deletePackage(element).subscribe(
-          (response: any) => {
-            this.toastr.success(response.message, '', {
-              timeOut: 3000,
-            });
-            this.loadPackageData();
-          },
-          (error) => {
-            this.toastr.error(error.message, '', {
-              timeOut: 3000,
-            });
-          }
-        );
-      }
-    });
+    const isIdInResponse = this.subscriptionList.some(
+      (item) => item.packageId._id === element._id
+    );
+
+    if (isIdInResponse) {
+      Swal.fire({
+        title: 'You cannot delete package in use',
+        icon: 'warning',
+        confirmButtonColor: '#3085d6',
+        confirmButtonText: 'OK'
+      });
+    } else {
+      Swal.fire({
+        title: 'Are you sure?',
+        text: 'You are about to delete this package.',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonText: 'Delete',
+        cancelButtonText: 'Cancel',
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.packageService.deletePackage(element).subscribe(
+            (response: any) => {
+              this.toastr.success(response.message, '', {
+                timeOut: 3000,
+              });
+              this.loadPackageData();
+            },
+            (error) => {
+              this.toastr.error(error.message, '', {
+                timeOut: 3000,
+              });
+            }
+          );
+        }
+      });
+    }
   }
 }
